requirements.txt
openai>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
numpy<2.0.0
python-dotenv